Fixed Roof controller UI issues
