+++ src/HVO.WebSite.RoofControllerV4/Components/Layout/MainLayout.razor
@@ -15,11 +15,6 @@
                     <div class="collapse navbar-collapse" id="navbarNav">
                         <ul class="navbar-nav ms-auto">
                             <li class="nav-item">
-                                <a class="nav-link" href="/roof-control">
-                                    <i class="bi bi-house-door"></i> Roof Control
-                                </a>
-                            </li>
-                            <li class="nav-item">
                                 <a class="nav-link" href="/roof-control-v2">
                                     <i class="bi bi-layout-wtf"></i> Roof Control V2
                                 </a>
