+++ src/HVO.WebSite.RoofControllerV4/Components/Pages/Home.razor
@@ -17,7 +17,7 @@
                         This system provides automated roof control and monitoring capabilities for the observatory.
                     </p>
                     <div class="d-grid gap-3 d-md-flex justify-content-md-center">
-                        <a class="btn btn-primary btn-lg" href="/roof-control" role="button">
+                        <a class="btn btn-primary btn-lg" href="/roof-control-v2" role="button">
                             <i class="bi bi-house-door me-2"></i>
                             Roof Control
                         </a>
