+++ src/HVO.WebSite.RoofControllerV4/Components/Pages/RoofControlV2.razor
@@ -1,9 +1,8 @@
 @page "/roof-control-v2"
-@inherits RoofControl
 @using HVO.WebSite.RoofControllerV4.Logic
 @using HVO.WebSite.RoofControllerV4.Models
 @using Microsoft.AspNetCore.Components.Web
-@using Microsoft.AspNetCore.Components
+@rendermode @(new InteractiveServerRenderMode(prerender: false))
 
 <PageTitle>Roof Control V2 - HVOv9</PageTitle>
