+++ src/HVO.WebSite.RoofControllerV4/HVO.WebSite.RoofControllerV4.csproj
@@ -23,10 +23,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Remove="Logic/RoofControllerServiceWithSimulatedEvents.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
     <Folder Include="Services/" />
   </ItemGroup>
