+++ src/HVO.WebSite.RoofControllerV4/appsettings.Development.json
@@ -27,7 +27,7 @@
     "LogLevel": {
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning",
-      "HVO.WebSite.RoofControllerV4.Components.Pages.RoofControl": "Information"
+      "HVO.WebSite.RoofControllerV4.Components.Pages.RoofControlV2": "Information"
     }
   },
