+++ src/HVO.WebSite.RoofControllerV4/appsettings.json
@@ -11,7 +11,7 @@
     "LogLevel": {
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning",
-      "HVO.WebSite.RoofControllerV4.Components.Pages.RoofControl": "Information"
+      "HVO.WebSite.RoofControllerV4.Components.Pages.RoofControlV2": "Information"
     }
   },
